+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Release Versions:
 ## Upcoming changes
 
 - fix(controllers): check for finite wrench values (#216)
+- feat(controllers): allow control type change after construction (#217)
 
 ## 5.2.3